refactor: dao interface 분리
